backend/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
SQLAlchemy==1.4.48
PyMySQL==1.1.0
python-dotenv==1.0.0
marshmallow==3.19.0
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
